fix world transofrm move bug
